Execute convolutional network somewhat correctly.
A convolution operator was implemented to execute something akin to
conv2d in pytorch. The dimensions and memory layout are ordered such
that the results correspond with pytorch execution. Everything is
hardcoded, even some magic numbers are thrown in for good measure.
Cleaning this mess up is going to be a fun past time activity for the
masochisticly inclined, which includes me apparently.